remove submissionTab
