+++ frontend/src/components/common/customTab.js
@@ -1,43 +1,51 @@
-
 import React from 'react';
 import PropTypes from 'prop-types';
 import Icon from 'material-ui/Icon';
 import { Tab } from 'material-ui/Tabs';
 import ReportProblemIcon from 'material-ui-icons/ReportProblem';
-
+import CheckCircle from 'material-ui-icons/CheckCircle';
 import { withStyles } from 'material-ui/styles';
 
 
-const styleSheet = theme => ({
-  rootPrimarySelected: {
-    color: theme.palette.secondary[500],
-  },
-  rootInheritSelected: {
-    color: theme.palette.secondary[500],
-  },
-  label: {
-    display: 'flex',
-    alignItems: 'center',
-    fontSize: theme.typography.fontSize,
-    whiteSpace: 'normal',
-    [theme.breakpoints.up('md')]: {
-      fontSize: theme.typography.fontSize - 1,
+const styleSheet = (theme) => {
+  const paddingIcon = theme.spacing.unit / 2;
+
+  return {
+    rootPrimarySelected: {
+      color: theme.palette.secondary[500],
     },
-  },
-});
+    rootInheritSelected: {
+      color: theme.palette.secondary[500],
+    },
+    checked: {
+      fill: '#189a58',
+      margin: `0 0 0 ${paddingIcon}px`,
+    },
+    label: {
+      display: 'flex',
+      alignItems: 'center',
+      fontSize: theme.typography.fontSize,
+      whiteSpace: 'normal',
+      [theme.breakpoints.up('md')]: {
+        fontSize: theme.typography.fontSize - 1,
+      },
+    },
+  };
+};
 
 const CustomTab = (props) => {
-  const { classes, warn, label, ...tabProps } = props;
+  const { classes, warn, checked, label, ...tabProps } = props;
   return (
     <Tab
       label={[
         label,
         warn && <Icon color="error"><ReportProblemIcon /></Icon>,
+        checked && <CheckCircle className={classes.checked} />,
       ]}
       classes={{
         rootPrimarySelected: classes.rootPrimarySelected,
         rootInheritSelected: classes.rootInheritSelected,
-        label: warn && classes.label,
+        label: (warn || checked) && classes.label,
       }}
       {...tabProps}
     />
@@ -49,6 +57,7 @@ CustomTab.propTypes = {
   label: PropTypes.string,
   iconStyle: PropTypes.string,
   warn: PropTypes.bool,
+  checked: PropTypes.bool,
 };
